+++ drape_frontend/visual_params.cpp
@@ -93,7 +93,7 @@ string const & VisualParams::GetResourcePostfix(double visualScale)
   // Looking for the nearest available scale.
   int postfixIndex = -1;
   double minValue = numeric_limits<double>::max();
-  for (int i = 0; i < ARRAY_SIZE(postfixes); i++)
+  for (int i = 0; i < static_cast<int>(ARRAY_SIZE(postfixes)); i++)
   {
     double val = fabs(postfixes[i].second - visualScale);
     if (val < minValue)
