+++ geometry/spline.cpp
@@ -15,9 +15,9 @@ Spline::Spline(vector<PointD> const & path)
   m_direction = vector<PointD>(cnt);
   m_length = vector<double>(cnt);
 
-  for(int i = 0; i < cnt; ++i)
+  for(size_t i = 0; i < cnt; ++i)
   {
-    m_direction[i] = path[i+1] - path[i];
+    m_direction[i] = path[i + 1] - path[i];
     m_length[i] = m_direction[i].Length();
     m_direction[i] = m_direction[i].Normalize();
   }
